# src/main/resources/schema/swagger/covid19-swagger-spec.yml
swagger: '2.0'
info:
  version: 0.0.1
  title:  Covid19-case-follow-up
  description: ''
host: 'localhost:8080'
basePath: /
securityDefinitions:
  basicAuth:
    type: basic
schemes:
  - http
...
paths:
...
  '/v1/register-follow-up':
      put:
        operationId: compileFollowUpQuestionnier
        summary: 'Register case follow-up forms'
        tags:
        - Case Follow up
        consumes:
        - application/json
        parameters:
          - name: body
            in: body
            required: true
            schema:
              $ref: '#/definitions/request-save-follow-up'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-base'
          '401':
            $ref: '#/responses/response-unauthorized'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
  '/v1/follow-up/{code}':
      get:
        operationId: getCaseFollowUpQuestionniers
        summary: Get follow up questionniers
        tags:
        - Case Follow up
        parameters:
          - $ref: '#/parameters/case-code-in-path'
        produces:
        - application/json
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-pui-follow-up-list'
          '401':
            $ref: '#/responses/response-unauthorized'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
parameters:
...
  case-code-in-path:
    in: path
    name: code
    type: string
    format: uuid
    required: true
...
definitions:
    response-base:
      type: object
      properties:
        success:
          type: boolean
        resultCode:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        transactionId:
          type: string
        type:
          type: string
      discriminator: type
      required:
        - success
        - resultCode
    model-case:
      type: object
      properties:
        id:
          type: integer
          format: int32
        caseCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum: 
          - F
          - M
        occupation:
          type: string
        passportNumber:
          type: string
        incidentContactPhone1:
          type: string
        incidentContactPhone2:
          type: string
        countryOfResidence:
          type: string
        countryOfOrigin:
          type: string
        nationality:
          type: string
        region:
          type: string
        subcityOrZone:
          type: string
        woreda:
          type: string
        kebele:
          type: string
        houseNo:
          type: string
        phoneNo:
          type: string
        latitude:
          type: string
        presumptiveResult:
          $ref: '#/definitions/model-enum-id-value'
        confirmedResult:
          $ref: '#/definitions/model-enum-id-value'
        identifiedBy:
          $ref: '#/definitions/model-enum-id-value'
        parentCaseCode:
          type: string
        recentTravelTo:
          type: string
        status:
          $ref: '#/definitions/model-enum-id-value'
        modifiedBy:
          type: string
        reportingDate:
          type: string
          format: date-time
        modifiedDate:
          type: string
          format: date-time
      required:
        - firstName
        - lastName
        - passportNumber
        - incidentContactPhone1
        - countryOfResidence
        - identifiedBy
        - modifiedBy
...
    model-pui-follow-up:
      type: object
      properties:
        puiCaseCode:
          type: string
        qId: 
          type: integer
          format: int32
        question:
          type: string
        selectedOption:
          type: string
        description:
          type: string
        modifiedBy:
          type: integer
          format: int32 
        insertDate:
          type: string
          format: date-time
      required:
        - puiCaseCode
        - qId
        - selectedOption
    model-enum-id-value:
      type: object
      properties:
        id: 
          type: integer
          format: int32
        value:
          type: string
...
    model-pui-follow-up-list:
      type: object
      properties:
        list:
          type: array
          items: 
            $ref: '#/definitions/model-pui-follow-up'
      required:
        - list
    request-save-case:
      allOf:
        - $ref: '#/definitions/model-case'
...
    request-save-follow-up:
      allOf:
        - $ref: '#/definitions/request-save-case'
        - $ref: '#/definitions/model-pui-follow-up-list'
...
    response-pui-follow-up-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-pui-follow-up-list'
responses:
    response-internal-server-error-500:
      description: Internal Server error
      schema:
        $ref: '#/definitions/response-base'
    response-unauthorized:
        description: Authentication information is missing or invalid
        schema:
          $ref: '#/definitions/response-base' 
        headers:
          WWW_Authenticate:
            type: string